% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{dyncli}
\alias{dyncli}
\title{\code{dyncli}: CLI Helper for Trajectory Inference Methods}
\description{
Helps with the input/output of trajectory inference methods created with \code{dynwrap}.
}
