% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{dyncli main function}
\usage{
main(args = commandArgs(trailingOnly = TRUE),
  definition_location = "/code/definition.yml")
}
\arguments{
\item{args}{The arguments to be read by dyncli}

\item{definition_location}{The location of the definition file of the method}
}
\description{
dyncli main function
}
