% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output}
\alias{write_output}
\title{Write TI output data to a file}
\usage{
write_output(x, task)
}
\arguments{
\item{x}{A dynwrap object or a list to be passed to \code{dynwrap::wrap_output_list()}}

\item{task}{The task read by \code{main()}}
}
\description{
Write TI output data to a file
}
